Add files via upload
Just a little piece of code that I worked hard on
I was tryna write a programm that inputes «character» variable and returns the output «OK/NOT OK» whether you've entered smth or not, BUT «space» symbol is considered as «NOT OK» too because technically you didn't really enter anything. 
So I got it done. I know it's small thing for someone but not for me. Peace✌