Add files via upload
It's a basic program that changes a singular noun to a plural one in English. I know there is <import inflect> in Python, but I decided to write my own program that does kind of the same thing. Of course that doesn't have a lot of exceptions in this language, but I'm gonna update this soon, but now here is what I got!